Code re-org
