+++ airflow_webserver/templates/airflow/dags.html
@@ -216,7 +216,7 @@
   <script src="{{ url_for('static', filename='bootstrap3-typeahead.min.js') }}"></script>
   <script>
 
-      const DAGS_INDEX = "{{url_for('HomeView.index')}}"
+      const DAGS_INDEX = "{{url_for('Airflow.index')}}"
       const ENTER_KEY_CODE = 13
 
       $('#dag_query').on('keypress', function (e) {
